Fix audited decompiler, CLI, parsing and packaging defects
Addresses the bulk of the findings recorded in AUDIT.md. Every fix below
was verified against the shipped jar, the compiled classes, or a real
Hammer-authored VMF; the full build stays green with 622 tests passing.

Decompiler output correctness:

- FaceSource: `i % 2 * psize` parses as `(i % 2) * psize`, so the row-flush
  condition was never true and triangle_tags was emitted empty for every
  displacement. Confirmed against test/maps/sdk_dm_lockdown.vmf, where
  triangle_tags is psize rows of 2*psize entries, matching the corrected
  grouping and DDispInfo.getTriangleTagCount().
- TextureSource: cubemaps were bound when any single coordinate matched
  rather than all three.
- EntitySource: overlay fade distances are stored squared and were written
  straight into fademindist/fademaxdist. Negative values are a "no fade"
  sentinel (the golden VMF uses -1), so they pass through unconverted.
- BrushSideFaceMapper: the winning original face was selected by largest
  index instead of by the overlap area the stream had just computed, and a
  0/0 percentage printed NaN in normal decompiles.
- VmfWriter: PrintWriter swallows IOException and checkError was never
  called, so a truncated VMF was reported as a successful decompile.
- BrushSource: the one-argument writeModel recursed into itself.
- AreaportalMapper: a portal assigned to slot 0 was treated as unassigned.
- VmfMeta: visgroup colors were written red-blue-green.

Game detection: nine file-name patterns could not match any real map name
of their own game. Six carried a `$` anchor on what is meant to be a prefix
(`^mp_$` never matches `mp_angel_city`), and three used a character class
where alternation was meant (`[de|cs]` is the set {d,e,|,c,s}, so the CS
patterns never matched `de_dust2`). Each corrected pattern was checked
against its own maps and against other games' names.

CLI: passing a directory matched zero maps because Files.walk used
maxDepth 0, and every failure path exited 0. The command now reports
failures through its exit code. The --brushmode help printed "null" four
times via unsupported interpolation, so no valid value was discoverable.

Parsing robustness: lump length checks overflowed int before being
clamped; an off-by-one accepted lump index 64; a failed LZMA decompression
still cleared the compressed flag; pakfile extraction wrote directory
entries as files and then aborted the whole archive; EntityIO read
timesToFire from the delay field.

bspsrc-common: CountingInputStream's bulk reads called the single-byte
delegate, never filled the caller's array and returned the byte value as
the count; its counter summed byte values. Collectors.mode's combiner
dropped one partial map. AlphanumComparator returned 0 for unequal
strings; fixing it in the chunk comparison rather than as a final tiebreak
also removes an intransitivity, verified over 17.5M triples.

Packaging and CI: the shaded jar excluded its own generated
Log4j2Plugins.dat, so the jar-only distribution logged an error on launch
and never installed its console filter; the descriptors are now merged.
Tests run on every push and pull request instead of only on release tags,
bspsrc-app declares the bspsrc-lib dependency its module-info requires,
and generated launcher scripts quote "$@" so paths with spaces survive.

+++ bspsrc-lib/src/main/java/info/ata4/bspsrc/lib/BspFile.java
@@ -328,7 +328,7 @@ public class BspFile {
             }
 
             // fix invalid lengths
-            if (ofs + len > bb.limit()) {
+            if ((long) ofs + (long) len > bb.limit()) {
                 int lenOld = len;
                 len = bb.limit() - ofs;
                 L.warn("Invalid lump length {} in {}, assuming {}", lenOld, ltype, len);
